refactor(console): extract api details danger-zone to dedicated component
