+++ gravitee-apim-console-webui/src/management/api/details/api-portal-details-danger-zone/api-portal-details-danger-zone.component.scss
@@ -1,0 +1,28 @@
+@use 'sass:map';
+@use '@angular/material' as mat;
+@use '@gravitee/ui-particles-angular' as gio;
+
+:host {
+  display: block;
+}
+
+.danger-card {
+  border: 1px solid mat.get-color-from-palette(gio.$mat-error-palette, 'default');
+
+  &__title {
+    color: mat.get-color-from-palette(gio.$mat-error-palette, 'default');
+  }
+
+  &__actions {
+    display: flex;
+    flex-direction: column;
+
+    &__action {
+      display: flex;
+      align-items: center;
+      flex-direction: row;
+      justify-content: space-between;
+      height: 48px;
+    }
+  }
+}
