+++ gravitee-apim-console-webui/src/management/api/details/api-portal-details.component.scss
@@ -115,24 +115,6 @@ $typography: map.get(gio.$mat-theme, typography);
   margin-top: 16px;
 }
 
-.danger-card {
+.danger-zone {
   margin-top: 16px;
-  border: 1px solid mat.get-color-from-palette(gio.$mat-error-palette, 'default');
-
-  &__title {
-    color: mat.get-color-from-palette(gio.$mat-error-palette, 'default');
-  }
-
-  &__actions {
-    display: flex;
-    flex-direction: column;
-
-    &__action {
-      display: flex;
-      align-items: center;
-      flex-direction: row;
-      justify-content: space-between;
-      height: 48px;
-    }
-  }
 }
